pyproject.toml
[tool.poetry]
name = "la2i-flask"
version = "0.1.0"
description = ""
authors = ["paisleypark <paisleypark@outlook.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
passlib = "^1.7.4"
jwt = "^1.3.1"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"
flask-session = "^0.5.0"
python-dotenv = "^1.0.0"
langchain = "^0.0.354"
youtube-transcript-api = "^0.6.2"
openai = "^1.6.1"
networkx = "^3.2.1"
matplotlib = "^3.8.2"
frontend = "^0.0.3"
tools = "^0.1.9"
pdfplumber = "^0.10.3"
chromadb = "^0.4.22"
tiktoken = "^0.5.2"
faiss-cpu = "^1.7.4"
pymupdf = "^1.23.17"
pyvis-network = "^0.0.6"
pyvis = "^0.3.2"
wikipedia = "^1.4.0"
latexcodec = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
